Move common field transformers to apps/transformers
